// back-end/apps/back-end/src/common/types/generate-contract.type.ts
export type GenerateContractImage = {
  lessorName: string;
  lessorAddress: string;
  lessorEmail: string;
  lesseeName: string;
  lesseeAddress: string;
  lesseeEmail: string;
  treeNames: string[];
  totalTree: number;
  farmAddress: string;
  startDate: string;
  endDate: string;
  totalPrice: number;
  currentDate: string;
  currentMonth: string;
  currentYear: string;
};